Move prepublish checks before building packages for publishing
