+++ package.json
@@ -13,8 +13,9 @@
         "prettier:ci": "prettier --list-different '**/*.{ts,tsx,json,md}' --config .prettierrc",
         "report_coverage": "lcov-result-merger 'packages/*/coverage/lcov.info' | coveralls",
         "test:installation": "node ./packages/monorepo-scripts/lib/test_installation.js",
-        "run:publish": "run-s install:all rebuild script:publish",
-        "run:publish:dry": "run-s install:all rebuild script:publish:dry",
+        "run:publish": "run-s install:all build:monorepo_scripts script:prepublish_checks rebuild script:publish",
+        "run:publish:dry": "run-s install:all build:monorepo_scripts script:prepublish_checks rebuild script:publish:dry",
+        "script:prepublish_checks": "node ./packages/monorepo-scripts/lib/prepublish_checks.js",
         "script:publish": "node ./packages/monorepo-scripts/lib/publish.js",
         "script:publish:dry": "IS_DRY_RUN=true yarn script:publish",
         "install:all": "yarn install",
@@ -22,6 +23,7 @@
         "lerna:run": "lerna run",
         "watch": "wsrun watch $PKG --fast-exit -r --stages --done-criteria='complete|successfully'",
         "build": "wsrun build $PKG --fast-exit -r --stages",
+        "build:monorepo_scripts": "PKG=@0xproject/monorepo-scripts yarn build",
         "clean": "wsrun clean $PKG --fast-exit -r --parallel",
         "rebuild": "run-s clean build",
         "test": "wsrun test $PKG --fast-exit --serial --exclude-missing",
